requirements.txt
asgiref==3.4.1
astroid==2.9.3
click==8.0.3
Django==4.0.1
django-framework==0.1.2
djangorestframework==3.13.1
Flask==2.0.2
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
Pillow==9.0.0
platformdirs==2.4.1
pylint==2.12.2
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
Werkzeug==2.0.2
wrapt==1.13.3
